Fix overflow on 64-bit with negative offset
When compressing `6fiffy4.wad` on 64-bit, the sidedefs end up larger
than the original lump (for documented reasons). In this scenario we are
supposed to show a negative count of bytes saved. However, because the
original values are `uint32_t`s, the value would overflow.

On 32-bit machines this behavior was likely masked as `long` was 32-bit
and the resulting value would become signed, but on 64-bit we were
instead getting a huge value. We can avoid the overflow by just casting
to `long` before performing the calculation.

+++ main.c
@@ -401,7 +401,8 @@ static bool TryPack(wad_file_t *wf, unsigned int lump_index, FILE *out_file,
                 PercentSmaller(orig_lump_len, wf->entries[lump_index].length));
             *sidedefs_larger = *sidedefs_larger ||
                                wf->entries[lump_index].length > orig_lump_len;
-            stats->packed += orig_lump_len - wf->entries[lump_index].length;
+            stats->packed += (long) orig_lump_len
+                           - (long) wf->entries[lump_index].length;
         }
         else
         {
@@ -437,7 +438,8 @@ static bool TryStack(wad_file_t *wf, unsigned int lump_index, FILE *out_file,
         SPAMMY_PRINTF(
             "(%s), done.\n",
             PercentSmaller(orig_lump_len, wf->entries[lump_index].length));
-        stats->stacked += orig_lump_len - wf->entries[lump_index].length;
+        stats->stacked += (long) orig_lump_len
+                        - (long) wf->entries[lump_index].length;
     }
     else
     {
@@ -472,7 +474,8 @@ static bool TrySquash(wad_file_t *wf, unsigned int lump_index, FILE *out_file,
     SPAMMY_PRINTF(
         "(%s), done.\n",
         PercentSmaller(orig_lump_len, wf->entries[lump_index].length));
-    stats->squashed += orig_lump_len - wf->entries[lump_index].length;
+    stats->squashed += (long) orig_lump_len
+                     - (long) wf->entries[lump_index].length;
 
     return true;
 }
